fix: fixed types
