--- src/waiting-time/entities/withdrawal-time-calculation-type.enum.ts
@@ -1,9 +1,0 @@
-export enum WaitingTimeCalculationType {
-  buffer = 'buffer',
-  bunker = 'bunker',
-  vaultsBalance = 'vaultsBalance',
-  rewardsOnly = 'rewardsOnly',
-  validatorBalances = 'validatorBalances',
-  requestTimestampMargin = 'requestTimestampMargin',
-  exitValidators = 'exitValidators',
-}
