+++ src/waiting-time/index.ts
@@ -1,4 +1,3 @@
 export * from './waiting-time.service';
 export * from './waiting-time.module';
 export * from './waiting-time.types';
-export * from './entities/withdrawal-time-calculation-type.enum';
